+++ src/server/include/nms_dcoll.h
@@ -272,7 +272,7 @@ protected:
    BYTE m_status;                // Item status: active, disabled or not supported
    BYTE m_busy;                  // 1 when item is queued for polling, 0 if not
 	BYTE m_scheduledForDeletion;  // 1 when item is scheduled for deletion, 0 if not
-	uint16_t m_flags;
+	uint32_t m_flags;
 	uint32_t m_dwTemplateId;         // Related template's id
 	uint32_t m_dwTemplateItemId;     // Related template item's id
    MUTEX m_hMutex;
